Add electron-builder config for linux desktop entry
This adds the association of the `earthdata-download://` protocol to linux via a
.desktop entry.

Adapted from: https://stackoverflow.com/questions/65808445/electron-how-to-create-deep-linking-on-linux

Note that this does ont resolve the issue with deep links not being associated
with the running app in dev (`npm start`). This does not appear to be supported
out-of-the-box.

https://www.electronjs.org/docs/latest/tutorial/launch-app-from-url-in-another-app#packaging

> On macOS and Linux, this feature will only work when your app is packaged. It
will not work when you're launching it in development from the command-line.